FLIZ-185 refactor: NotificationSearch 구성 컴포넌트를 개별 파일로 분리
